Add portfolio table with mock data (WIP)
