+++ src/constants/colors.ts
@@ -23,4 +23,5 @@ export const colors = {
     800: 'rgba(0, 0, 0, 0.80)',
     900: 'rgba(0, 0, 0, 0.92)',
   },
+  "table-row": '#F5F9FA'
 };
