+++ src/pages/bridge/Transfers.tsx
@@ -150,7 +150,7 @@ function Transfers(): JSX.Element {
         />
       )}
       <Table
-        className="transfer-list-table bg-base-100 text-md"
+        className="transfer-list-table bg-base-100 text-md w-full"
         data={data}
         columns={columns}
         isLoading={false}
